+++ tom_d4rt/lib/src/stdlib/convert/latin1.dart
@@ -71,14 +71,13 @@ class Latin1EncoderConvert {
       },
       'startChunkedConversion':
           (visitor, target, positionalArgs, namedArgs, _) {
-            if (positionalArgs.length != 1 ||
-                positionalArgs[0] is! Sink<List<int>>) {
+            if (positionalArgs.length != 1 || positionalArgs[0] is! Sink) {
               throw RuntimeD4rtException(
                 'startChunkedConversion requires a Sink<List<int>> argument.',
               );
             }
             return (target as Latin1Encoder).startChunkedConversion(
-              positionalArgs[0] as Sink<List<int>>,
+              D4.adaptSink<List<int>>(positionalArgs[0], 'sink'),
             );
           },
       'bind': (visitor, target, positionalArgs, namedArgs, _) {
@@ -131,14 +130,13 @@ class Latin1DecoderConvert {
       },
       'startChunkedConversion':
           (visitor, target, positionalArgs, namedArgs, _) {
-            if (positionalArgs.length != 1 ||
-                positionalArgs[0] is! Sink<String>) {
+            if (positionalArgs.length != 1 || positionalArgs[0] is! Sink) {
               throw RuntimeD4rtException(
                 'startChunkedConversion requires a Sink<String> argument.',
               );
             }
             return (target as Latin1Decoder).startChunkedConversion(
-              positionalArgs[0] as Sink<String>,
+              D4.adaptSink<String>(positionalArgs[0], 'sink'),
             );
           },
       'bind': (visitor, target, positionalArgs, namedArgs, _) {